requirements.txt
asgiref==3.11.1
attrs==25.4.0
autobahn==25.12.2
Automat==25.4.16
cbor2==5.8.0
cffi==2.0.0
channels==4.1.0
constantly==23.10.4
cryptography==46.0.5
daphne==4.1.2
Django==5.2.11
djangorestframework==3.16.1
et_xmlfile==2.0.0
fire==0.7.1
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
msgpack==1.1.2
openpyxl==3.1.5
packaging==26.0
pillow==12.1.1
prettytable==3.17.0
py-ubjson==0.16.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pyOpenSSL==26.0.0
pyzkaccess==1.1
service-identity==24.2.0
sqlparse==0.5.5
termcolor==3.3.0
Twisted==25.5.0
txaio==25.12.2
typing_extensions==4.15.0
tzdata==2025.3
ujson==5.12.0
watchdog==6.0.0
wcwidth==0.6.0
wrapt==1.17.3
zope.interface==8.2
